Add check for deb file

+++ setup/apm_planner_setup.sh
@@ -1,7 +1,10 @@
 #!/usr/bin/env bash
 
 cd ~/Downloads
-wget http://ardupilot.com/wp-content/plugins/download-monitor/download.php?id=111 -O apmplanner2.deb
+if [! -f ~/Download/apmplanner2.deb ]; then
+    wget http://ardupilot.com/wp-content/plugins/download-monitor/download.php?id=111 -O apmplanner2.deb
+fi
+
 sudo dpkg -i apmplanner2.deb
 sudo apt-get -f install
 sudo dpkg -i apmplanner2.deb
